chore: update go mod to remove requires
This PR updates the go mod to use the latest v0.3 capi release, as well
as downgrade controller-runtime to match. The only real fallout of this
was the removal of the ctx from some calls to the k8s api, so shouldn't
be a big deal to readd this stuff once we move to v0.4 of capi.

Signed-off-by: Spencer Smith <spencer.smith@talos-systems.com>